+++ app/src/main/java/tesis/untref/com/firealerts/model/Latitude.kt
@@ -1,3 +1,3 @@
 package tesis.untref.com.firealerts.model
 
-data class Latitude (val degree: Int, val minute: Int, val second: Float, val cardinalPoint: CardinalPoint)
+data class Latitude (val degree: Int, val minute: Int, val second: Double, val cardinalPoint: CardinalPoint)
