+++ app/src/main/java/tesis/untref/com/firealerts/model/Longitude.kt
@@ -1,3 +1,3 @@
 package tesis.untref.com.firealerts.model
 
-data class Longitude (val degree: Int, val minute: Int, val second: Float, val cardinalPoint: CardinalPoint)
+data class Longitude (val degree: Int, val minute: Int, val second: Double, val cardinalPoint: CardinalPoint)
